iss #97 there were three different plot_difference functions in this repository so I am condensing them into two functions in one file
